ValidationKit/utils/IoPerf: Start on a I/O benchmarking/testing utility for our automated tests using the new I/O queue API [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@80032 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/ValidationKit/utils/storage/IoPerf.cpp
@@ -236,7 +236,7 @@ static uint64_t     g_cbTestSet    = _2G;
 /** Block size for each request. */
 static size_t       g_cbIoBlock    = _4K;
 /** Maximum number of concurrent requests for each job. */
-static size_t       g_cReqsMax     = 16;
+static uint32_t     g_cReqsMax     = 16;
 
 
 /** @name Configured tests, this must match the IOPERFTEST order.
